Use GetModelOptions in foundry_get_model
